+++ client/components/mma/cancel/cancellationContributionReminder.tsx
@@ -4,12 +4,12 @@ import { Button, Radio, RadioGroup } from '@guardian/source-react-components';
 import { useEffect, useState } from 'react';
 import type * as React from 'react';
 import { useNavigate } from 'react-router';
-import { trackEventInOphanOnly } from '../../../utilities/analytics';
-import { getGeoLocation } from '../../../utilities/geolocation';
 import {
 	buttonCentredCss,
 	stackedButtonLayoutCss,
-} from './cancellationSaves/SaveStyles';
+} from '../../../styles/ButtonStyles';
+import { trackEventInOphanOnly } from '../../../utilities/analytics';
+import { getGeoLocation } from '../../../utilities/geolocation';
 
 const containerStyles = css`
 	padding-bottom: ${space[24]}px;
