+++ client/components/mma/cancel/cancellationSaves/SaveStyles.ts
@@ -87,67 +87,6 @@ export const listWithDividersCss = css`
 	}
 `;
 
-export const buttonCentredCss = css`
-	justify-content: center;
-`;
-
-export const buttonLayoutCss = css`
-	> * + * {
-		margin-left: ${space[3]}px;
-	}
-`;
-
-export const stackedButtonLayoutCss = css`
-	display: flex;
-	flex-direction: column;
-	margin-top: ${space[5]}px;
-	> * + * {
-		margin-top: ${space[3]}px;
-	}
-	${from.tablet} {
-		flex-direction: row;
-		> * + * {
-			margin-top: 0;
-			margin-left: ${space[3]}px;
-		}
-	}
-`;
-
-export const reverseStackedButtonLayoutCss = css`
-	display: flex;
-	flex-direction: column-reverse;
-	margin-top: ${space[5]}px;
-	padding-top: 32px;
-	> * + * {
-		margin-bottom: ${space[3]}px;
-	}
-	${from.tablet} {
-		flex-direction: row;
-		justify-content: flex-end;
-		> * + * {
-			margin-top: 0;
-			margin-left: ${space[3]}px;
-		}
-	}
-`;
-
-export const stackedButtonLeftLayoutCss = css`
-	display: flex;
-	flex-direction: column-reverse;
-	margin-top: ${space[5]}px;
-	padding-top: 32px;
-	> * + * {
-		margin-bottom: ${space[3]}px;
-	}
-	${from.tablet} {
-		flex-direction: row;
-		> * + * {
-			margin-top: 0;
-			margin-left: ${space[3]}px;
-		}
-	}
-`;
-
 export const smallPrintCss = css`
 	${textSans.xxsmall()};
 	margin-top: 0;
@@ -187,26 +126,6 @@ export const headingCss = css`
 	span {
 		display: block;
 		color: ${palette.brand['500']};
-	}
-`;
-
-export const buttonContainerCss = css`
-	${until.tablet} {
-		display: flex;
-		flex-direction: column;
-	}
-`;
-
-export const buttonMutedCss = css`
-	${until.tablet} {
-		border: none;
-	}
-`;
-
-export const wideButtonCss = css`
-	${from.tablet} {
-		flex-grow: 1;
-		max-width: 300px;
 	}
 `;
