+++ client/components/mma/cancel/cancellationSaves/SelectReason.tsx
@@ -23,6 +23,11 @@ import {
 	MDA_TEST_USER_HEADER,
 } from '../../../../../shared/productResponse';
 import type { ProductTypeWithCancellationFlow } from '../../../../../shared/productTypes';
+import {
+	buttonCentredCss,
+	stackedButtonLayoutCss,
+	wideButtonCss,
+} from '../../../../styles/ButtonStyles';
 import { GenericErrorScreen } from '../../../shared/GenericErrorScreen';
 import { JsonResponseHandler } from '../../shared/asyncComponents/DefaultApiResponseHandler';
 import type {
@@ -32,14 +37,7 @@ import type {
 import { CancellationContext } from '../CancellationContainer';
 import type { CancellationReason } from '../cancellationReason';
 import { membershipCancellationReasons } from '../membership/MembershipCancellationReasons';
-import {
-	buttonCentredCss,
-	headingCss,
-	paragraphListCss,
-	sectionSpacing,
-	stackedButtonLayoutCss,
-	wideButtonCss,
-} from './SaveStyles';
+import { headingCss, paragraphListCss, sectionSpacing } from './SaveStyles';
 
 const reasonLegendCss = css`
 	display: block;
