+++ client/components/mma/cancel/cancellationSaves/ValueOfSupport.tsx
@@ -4,17 +4,17 @@ import { Button, Stack } from '@guardian/source-react-components';
 import { useContext } from 'react';
 import { Navigate, useLocation, useNavigate } from 'react-router';
 import { dateString } from '../../../../../shared/dates';
+import {
+	buttonCentredCss,
+	reverseStackedButtonLayoutCss,
+} from '../../../../styles/ButtonStyles';
 import { ProgressStepper } from '../../shared/ProgressStepper';
 import type {
 	CancellationContextInterface,
 	CancellationRouterState,
 } from '../CancellationContainer';
 import { CancellationContext } from '../CancellationContainer';
-import {
-	buttonCentredCss,
-	headingCss,
-	reverseStackedButtonLayoutCss,
-} from './SaveStyles';
+import { headingCss } from './SaveStyles';
 
 export const ValueOfSupport = () => {
 	const navigate = useNavigate();
