+++ client/components/mma/switch/review/SwitchReview.tsx
@@ -14,6 +14,10 @@ import { Link } from 'react-router-dom';
 import { dateString } from '../../../../../shared/dates';
 import type { ProductSwitchType } from '../../../../../shared/productSwitchTypes';
 import {
+	buttonCentredCss,
+	buttonMutedCss,
+} from '../../../../styles/ButtonStyles';
+import {
 	LoadingState,
 	useAsyncLoader,
 } from '../../../../utilities/hooks/useAsyncLoader';
@@ -32,8 +36,6 @@ import type {
 } from '../SwitchContainer';
 import { SwitchContext } from '../SwitchContainer';
 import {
-	buttonCentredCss,
-	buttonMutedCss,
 	errorSummaryBlockLinkCss,
 	errorSummaryLinkCss,
 	errorSummaryOverrideCss,
